Fix another pointer to integer size mismatch on WIN64/LLP64
1) struct NSFastEnumerationState.state is also used to store (64 bit) pointers in WIN64/LLP64 so must be 64 bits wide.

+++ Headers/Foundation/NSEnumerator.h
@@ -39,7 +39,7 @@ extern "C" {
 
 typedef struct
 {
-  unsigned long	state;
+  uintptr_t	state;
   __unsafe_unretained id		*itemsPtr;
   unsigned long	*mutationsPtr;
   uintptr_t	extra[5];
